Add loading state provider
Add possibility to replace the loading state with a custom implementation. The goal is to have the same loading style as the other parts of the app.
Exposes the LoadingStateProvider with two simple methods to show and hide the loading.
Adds the backward compatible implementation and use it as default loading provider.